+++ scheduler/Client/Commons/trace.py
@@ -68,44 +68,45 @@ class TraceLog:
 
     # Visualiza hasta 100 eventos para cada hilo (si se desean más eventos por hilo debe enviarse por parámetro)
     def plot_events(self, event_per_thread=100):
-
+        '''
         print(" Event List before start")
         for i in range(len(self.events_list)-1):
             print(self.events_list[i].get_event_type())
+        '''
 
         data_plot=[[]]
         
         event_list_aux=[]
 
-        print(self.events_list)
+        #print(self.events_list)
 
-        for i in range(0, self.threads_count-1):
+        for i in range(self.threads_count):
             data_event=[]
             if (len(self.events_list)!=0):
                 index=0
                 for j in range(0, len(self.events_list)-1):
-                    print(" Event List before remove")
-                    for h in range(len(self.events_list)-1):
-                        print(self.events_list[h].get_event_type())
+                    #print(" Event List before remove")
+                    #for h in range(len(self.events_list)-1):
+                        #print(self.events_list[h].get_event_type())
                     if self.events_list[index].get_thread() == i:
                         data_event.append(self.events_list[index].get_time())
                         event_list_aux.append(self.events_list[index])
                         self.events_list.remove(self.events_list[index])
                     else:
                         index= index+1
-            print("Thread:" + str(i))
-            print("Data plot:")
-            print(data_plot)
+            #print("Thread:" + str(i))
+            #print("Data plot:")
+            #print(data_plot)
 
-            print("Data Event before fill to zeros:")
-            print(data_event)
+            #print("Data Event before fill to zeros:")
+            #print(data_event)
 
             # Completar el vector de datos del hilos para que todas las filas tenga la misma dimension
             while (len(data_event) < event_per_thread):
                 data_event.append(0)
 
-            print("Data Event after fill to zeros:")
-            print(data_event)
+            #print("Data Event after fill to zeros:")
+            #print(data_event)
 
             data_event= [data_event]
 
@@ -115,21 +116,21 @@ class TraceLog:
                 data_plot= np.concatenate((data_plot, data_event))
         self.events_list= event_list_aux
 
-        print("Data Plot final:")
-        print(data_plot)
+        #print("Data Plot final:")
+        #print(data_plot)
 
         # set different colors for each set of positions
         colors = ['C{}'.format(i) for i in range(self.threads_count)]
-        print("Colors of events:")
-        print(colors)
+        #print("Colors of events:")
+        #print(colors)
 
-        threads_plot=[i for i in range(0, self.threads_count)]
-        print("Event_plot:")
-        print(threads_plot)
+        threads_plot=[i for i in range(self.threads_count)]
+        #print("Event_plot:")
+        #print(threads_plot)
 
         line_lengths=[1 for i in range(self.threads_count)]
-        print("Line Lengths:")
-        print(line_lengths)
+        #print("Line Lengths:")
+        #print(line_lengths)
 
         # create a horizontal plot
         plt.eventplot(data_plot, colors=colors, lineoffsets=threads_plot, linelengths=line_lengths)
@@ -139,7 +140,7 @@ class TraceLog:
         plt.ylabel('events[integer]')
 
         for i in range(0, len(self.events_list)):
-            data_label=str(self.events_list[i].get_event_type())+"\n"+str(self.events_list[i].get_value()+"\n"+self.events_list[i].get_value2())
+            data_label=str(self.events_list[i].get_event_type())+"\n"+str(self.events_list[i].get_value())+"\n"+str(self.events_list[i].get_value2())
             plt.annotate(data_label,(self.events_list[i].get_time(), self.events_list[i].get_thread()))
 
         plt.show()